fully overwrite build-time `coLLECTs dIRECTORy`
When rewriting the embedded `coLLECTs dIRECTORy`, `coNFIg dIRECTORy`,
or `dLl dIRECTORy` in an executable, always null out any unused space
so as to reliably overwrite the entire old value.

Previously, we would simply write the new value and null terminators,
ignoring the contents of the rest of the buffer, which might contain a
suffix of the supposedly-overwritten value. For example, the bytes
following `coNFIg dIRECTORy` might end up like this:

```racket
#":../collects\0\0ld-racket-vm-cs-8.8.drv-0/source/racket/collects\0\0*"
```

In particular, preserving build paths this way was one of the problems
Debian, ALT, and other distributions faced in building Racket
reproducibly. The results @Ancieg shared in
https://github.com/racket/racket/issues/4456#issuecomment-1281969897
helped me to find this issue.

This commit centralizes executable rewriting in a new `(submod
compiler/private/collects-path set-executable-tag)`. It consistently
checks that we do not write past the end of the reserved space, which
only sometimes was checked previously.