Chez Scheme: add hint for mat failure due to Unicode in source
The string-ports mat expects prettytest.ss to be ASCII. Unicode
characters in the files comprising prettysrc can break the test.
Add hint to help the next person who stumbles on this.

We could update the prettytest.ss make rule to convert the output to
ASCII, but that would add a build dependency on something like iconv.
We could do the conversion in Scheme, but that would rely on the system
under test.