extend `:do-in` to allow an inner definition sequence
The `:do-in` form has always had a form allowed between the `outer-id`
bindings and other forms. The original intend was that a side-effect
check would go there, but it turns out that definitions are allowed
there (inside a `begin`, because it really does end up in a
`let-values` body. Allowing definitions turns out to be useful for
Rhombus, so that it can put `define-syntaxes` there. This commit adds
a similar, optional position after the `inner-id` bindings in
`:do-in`, which is useful to Rhombus for the same reason.