raco setup: handle ".dep" without ".zo" better
You'd expect that `raco setup` would always build ".zo" files when
they're not present. However, to support recompilation of
independent-files to machine-specific files and deal with various
locking constraints, it's possible for ".zo" creation to be skipped
(for a single `raco setup` run, at least) if its corresponding ".dep"
file exists. The rule has been to make sure you delete ".dep" files
when deleting ".zo" files --- but that's not obvious or easy to
remember.

The tweak here narrows the problem to ".zo" files that originated from
recompilation of machine-independent files, since that's the situation
that needs to accomodate ".zo" files that are temporarily missing.
That leaves the problem to just packages that are included in a
distribution or that are installed from a built-package catalog.

A more general solution needs a different locking strategy and API.