backward-compatible `syntax/unsafe/for-transform`
The `syntax/unsafe/for-transform` module exposes the internal form of
an expanded `for` clause, and that internal format changed in
b8dbadec94.

Adjust `expand-for-clause` so that it produces only the old format ---
which preserves the behavior of the `generic-bind` package, for
example. If a `for` clause uses the new `inner-check` position in
`:do-in`, then `expand-for-clause` reports an error. A new
`expand-for-clause*` supports the expanded `:do-in` format.