repair `module-path?` for trailing dots in directory elements
The `module-path?` parser did not track dots correctly, and a path
like `"x./y"` or `'racket...` was allowed.

This is a backward-incompatible change, but it seems unlikely that
anyone intends to use trailing dots like that in a directory or
collection name, and it brings `module-path?` in line with the
documentation.