makefiles: adjust `make` in `--prefix` mode
When configured with `--prefix`, change `make` so that it performs
the ".zo"- and documentation-build work (i.e., `raco setup` step) that
`make install` currently performs --- but confined to the build
directory, and then the build-staged data is merged into place with
`make install`. That way, `make install` can be relatively fast, and
the ".zo"- and documentation-build step gets the advantage of being
incremental like the rest of `make`.

This change only applies to `make` as run within a "src" build
directory, so it does not affect `make` in the repo's root directory.
It is also constrained to non-cross mode, for now, but that
restriction could be lifted with a little more work.

A new `info-domain-root` configuration option is needed to keep the
info-domain step of `raco setup` confined to the build directory.