logging: never send 'none events
The 'none event level was originally intended for specifying receivers
and not for logging messages, and its handling when sending was
inconsistent. Ideally, 'none would not be allowed as a level for
sending a message; for a kind of backward capability, it's allowed
here while messages at that level are consistently suppressed.

It's possible that some existing code is using 'none as a really a
level beyond 'fatal. But probably and hopefully not, since it would
not work consistently.

Closes #4778