CS: repair `char-foldcase` for some single-character cases
`(char-foldcase #\u1E9E)` should be `#\uDF` (in contrast to the
`(string-foldcase (string #\u1E9E))` result of "ss"). The character
foldings that that are different here correspond to cases that are "S"
status in "CaseFolding.txt", and the changes bring Racket CS in line
with Racket BC on those examples.

Thanks to Marc Feeley for noticing and reporting the bug.