CS: repair for long-running atomic callback
The main repair is to clean a pending check-for-thread-swap action
that can get installed by a long-funning foreign callback, such as one
that is parsing a PNG or JPEG file. The action is otherwise triggered
by the end of atomic mode, which can lead to a crash if that happens
within the scheduler (if it redundantly declares entry into and out of
atomic mode) instead of in a thread context.

The fact that atomic mode for a callback ends in a special way has
some implications for scheduling and thread breaks. This commit
includes an update to the `foreign/unsafe` manual to document the
thread-break limitation; the behavior is documented, insead of
changed, because changing would be relatively complex and expansive,
and an atomic callback probably shouldn't try to break its own thread,
anyway. The scheduling limitation could cause a Racket thread to get
too much time, but that's unlikely to create behavior outside the
normal variation of Racket thread scheduling.