net/sendurl: refactor to avoid "ill-typed" computation
browser can be either a symbol or false or a pair. When it's not a
symbol, doing assq browser in existing-browsers->exes will always
result in #f. While that is technically correct,
a better way would be to avoid this "ill-typed" computation entirely
and only do assq browser in existing-browsers->exes when we are sure
browser is a symbol.