doc: elaborate browser-preference? and add a command caveat
The caveat states that when a browser preference is a pair of strings,
users must keep `escape?` to be true. Previously, the doc states
(in the doc of `external-browser`) that functions in net/sendurl
will ensure that the URL is already encoded, which is false.
It is possible to set `escape?` to `#f` to avoid the escaping.