schemify: more systematic unwrapping of vacuous `let-values`
There are several places in `schemify` that appears to unwrap
*vacuous* `let-values` (i.e., without bindings) and friends.  This
commit adds an `unwrap-let` function and attempts to systematically
use it.  This way, more optimization can be done in `schemify`.  The
most apparent benefit is that cross-module inlining can work better in
the presence of vacuous `let-values`, possibly introduced by macros.