improve terminal-mode status information
The workers can be in three different states: productively making zos,
blocked waiting for a fellow worker to finish the zo they also need
(but some other worker got to first), or idle. Track these three and
send changes to the lines in terminal for all three. When we're not in
terminal mode, however, just print out when a worker starts in on a
new collection (but ignoring stolen work).

This plan seems to match what was happening originally but gives the
information needed to show more when we're using terminal controls