list doc: use defproc* to separate the improper list case
Prior this PR, the improper list case is lumped into the list case
(often via `or/c` or by absorbing the more specific contract into
`any/c`). This results in many counterintuitive contracts. E.g.,
`drop` has the contract `(-> any/c natural? any/c)`, even though
in most cases people use it in the way that complies
`(-> list? natural? list?)`.

This PR uses `defproc*` to separate the two cases, making the contracts
more precise. This style is not new: the documentation for the `append`
function for example is already following the style. The PR simply
applies the style more universally.

Also fixes the contract of `findf`, which actually accepts
an improper list.