raco make: fix update of timestamps on ".zo" files
When a file's timestamp changes but the content doesn't since ".zo"
files were built, `raco make` (or a use of `compiler/cm` more
generally) is meant to update the ".zo" file's timestamp. That wasn't
happening for files that are `require`d by a file whose timestamp is
updated.

Compilation from machine-independent to machine-dependent form needs a
similar transitive operation, and updating timestamps now imitates and
simplifies that handling. The simplication is that when a module file
in a mode that cannot compile it (if needed), then there's no attempt
to avoid another check later in compile-enabled mode. That change can
lead to a second check of an up-to-date file that wouldn't have
happened before, but enough information is cached so that a
potentially redundant check is cheap.