fix `for/fold` binding in the presence of `#:result`
This bug was introduced in 3b5b0a7.  The `fold-var`s incorrectly
scoped over the whole `for/fold` expansion, while the "remaining
weirdness" behavior was that outermost iteration clauses should not
see the `fold-var`s.

Also revert the `for/foldr` change that appears to be incorrect.
Unlike `for/fold`, `for/foldr` does not need to use this approach,
because `fold-var`s are only visible to the body.

Related to #4898.

(cherry picked from commit a220d4da4d846a50f4da93a9771e277241cdaac9)