change ".zo" form to have per-target cross-module info
Previously, cross-module optimization information attached to a module
was required to be machine-independent, because compiling on machine C
for target T would use the information from an imported module loaded
for C to compile a new module for T. That constraint has become
limitng, particularly for static generation of foreign-function code.

This commit changes cross-module information (for CS) to be specific
to a machine, and it adds `compiled-expression-add-target-machine` so
that information from a compilation for T can be added to the compiled
form for C. The `compile/cm` process for cross-compilation uses the
new function to preserve cross-module optimization in
cross-compilation mode.