acks: add a check to make sure names are sorted
`raco test acks.rkt` checks that the list is properly sorted.
If it is not sorted, one can run `racket acks.rkt` to output
the sorted list expression, that can then be used to make
the test passed.

Last name is automatically inferred, but in case where the inference is
wrong, we can manually mark the last name with "@". An example from the
test: the inferred last name from "Andries van Dam" is "Dam", but
"Andries @van Dam" would mark that the last name is "van Dam".