avoid non-determinism in module scopes
Conceptually, a module doesn't just have just one scope, but one scope
for every possible phase. That's implemented internally through a
"multi-scope" object to contain phase-specific scopes that are created
on demand. Those phase-specific scopes have so far received an
identity based on when they're needed, which makes relative order
sensitive to the order that some things happen, including resolving
bulk imports. It also seems possible that a fresh scope would not
properly reflect its conceptual age (which is based on ordering of
ids) when it is created on demand.

To fix this, exploit the fact that a scope's identity has internally
been an exact integer, so a phase-specific scope can use an exact
fractional offset from the multi-scope's identity with the fraction
determined by the phase.

This change is intended to fix the deterministic-zo test on
"setup/private/setup-fprintf.rkt", where the `test` submodule creates
a bulk import, and the timing to flatten that import internally can
depend on expansions that have happened before.