change `chaperone-treelist` to support state-passing and more operations
Thread a "state" value through interposition procedures for
`chaperone-treelist` and family. On each operation that produces an
updated treelist, the updated state is associated with the chaperoned
result, so that interposition procedures can track the evolution of
the wrapped treelist as needed (with no dependence on mutable state).

Previously, operations that remove elements for a treelist had no
interposition procedures, but new ones have been added so that they
get the opportunity to update the state.

This is a backward-incomptible change within the development sequence,
but on the grounds that treelists are new and experimental, anyway.