CS: improve hash code for immutable hash tables
The old hash-code function for immutable hash tables tried using the
mask in HAMT nodes as a summary and substitute for key hashes. That's
not good enough, though, as illustrated by the case of making
different hash tables all with 1 key (so there are only 16 possible
masks in the sole node).

Closes #4902