guide: update other-editors for new Vim versions
Now that Vim is including some of my plugin, let's mention that the out
of the box experience for the racket filetype is alright. It's still
worth installing my plugin if you like to live on the edge (or want the
improved filetype detection).

While we're here, tidy up a few other passages.