guide: simplify Vim documentation
This version addresses feedback that the Vim guide is overly technical
by
- focusing on the provided features and recommended plugins early, and
- deferring discussion of minutiae like older versions and plugin
  managers to later.

In addition, it streamlines many documentation sections to focus on the
essentials and leave details for plugin documentation or other sites. At
the risk of being self-serving, it also cuts mention of an unmaintained
plugin in favor of mine (parts of which are now distributed with Vim).

One point of confusion likely remains: Scheme vs. Racket. I have not yet
contributed my enhanced filetype detection to Vim: it's a large change
from the current experience (if you're not using my plugin) and warrants
some care. Therefore this guide continues to point out the distinction
between scheme and racket filetypes and recommends my plugin for
enhanced support wherever possible. Configuring your own filetype
detection for the racket filetype to benefit from the out-of-the-box Vim
support continues to be an option.