add entitlement support to codesign implementation
The immeidate motivation for this change is to enable adding a
`com.apple.security.get-task-allow` entitlement to a build to so that
core dumps can work. Although Racket disbutions are signed with
entitlements, the functionality here is not needed for that step,
since the `codesign` tool is used for distributions.