add internal `unsafe-impersonate-hash`
Unlike `impersonate-hash`, `unsafe-impersonate-hash` works on
immutable hash tables. As a result, interposition procedures on hash
table operations can implement a custom hashing strategy. It's unsafe
because impersonating an immutable data structure can be misleading,
and the new operation is intended for use only when the impersonated
hash table is not itself visible. For now, this function is "internal"
in the sense that it's not documented and it's exported only by the
predefined `#%unsafe` module; it may change or go away in future
versions.