repair and performance improvement for `require` traversals
The `namespace-attach-module` function did not work right for a module
dependency that runs code at a phase level shifted to phase 0 while
the module's instantiation phase is non-0.

Internally, `require` module-name resolutions are cached differently
so that they can be shared as modules are attached. This change can
greatly reduce calls to the module name resolver when creating many
sandboxes for the same language, such as when rendering documentation.

Unnecessary traversal of module requires is now skipped in the case of
visiting available modules. In addition, a new `#:flatten-requires`
option in the `#%declare` form provides a hint that it's worth
flattening the tree of requires within a module's compiled form. This
flattening is now always a win; if two modules that have flattened
`require`s share a large subtree, and if those two modules are
required by a third, then instantiating the third can be slower than
without flattening. Flattening sparingly at the level of modules that
are often attached to oter namespaces, however, is likely to be
worthwhile. The `racket/base` module is marked `#:flatten-requires`,
and that helps make `(make-base-namespace)` twice as fast as before.