raco test: set current-test-invocation-directory
Running "raco test foo/bar/test.rkt" will cause Racket to change
directory to "foo/bar" before invoking "test.rkt", and error messages
are printed relative to (current-directory) like

    raco test: "foo/bar/test.rkt"
    --------------------
    fail
    FAILURE
    name:       check-equal?
    location:   test.rkt:5:0
    actual:     1
    expected:   2
    --------------------
    1/1 test failures

For an editor to consume this, it has to post-process the output to
correlate "raco test: path" with "location: partial path" (which is
difficult when running tests in parallel). This is not ideal; output
with

    location:   foo/bar/test.rkt:5:0

is immediately useable by editors for jumping to the error.

Expose the original directory in which "raco test" was launched so
that rackunit (or other test frameworks) can appropriately adjust
filepaths when reporting errors.

Also add a name to the test-log-enabled? parameter while we're here.