raco test: set more current-test-invocation-directory
This adds the effects of df01b0d1dc (raco test: set
current-test-invocation-directory, 2024-05-07) to --place and --process
mode, too; as a result, parallel invocations with -j work, too. To pass
the path to a subprocess, we need to convert to bytes (since paths do
not round trip through strings).

The changes to the subprocess argv processing and place channel received
list processing do perhaps beg the question: is the protocol for
unpacking the values communicated via side-channel error prone?