configure: base default package directory on `datadir`, not `datarootdir`
The default directory location should depend on `datadir` like the
location of the "collects" directory does, while `datarootdir` is for
system-recognized resources like man pages and program launchers.

There's some chance that this change can break or change the behavior
of existing install scripts, but hopefully not, since `datadir`
defaults to `datarootdir`.

Related to #4999