more consistently apply `'definition-intended-as-local` property
Clean up `racket/splicing` to more consistently apply the
`'definition-intended-as-local` syntax property for every splicing
form through the `splicing-let-start/def` internal form, which
fixes #4993 and a similar bug for `splicing-letrec-syntaxes+values`.
In addition, fix multiple body forms in
`splicing-letrec-syntaxes+values`.

Add test cases for `splicing-letrec-syntaxes+values` and
`splicing-local` (which have different implementations than other
splicing forms), and update an old test case that did not test what it
was supposed to test.

Update the doc to index the `'definition-intended-as-local` syntax
property.