add `compiler/demod`, support submodules via `raco demod`
Using `#lang compiler/demod` performs the same kind of module
flattening as `raco demod`, but in a way that cooperates with tools
like `raco make` and `raco setup`. For demodularizing a library, as
opposed to a program for end users, a module language is much easier
to use than a separate `raco demod` tool.

For example, the following "alamgam.rkt" module has the same provided
names and submodules as "amalgam-src.rkt", but it has its own copy of
any transitively referenced module within the "amalgam" subdirectory:

```
 ;; alamgam.rkt
 #lang racket/demod
 "amalgam-src.rkt"
 #:include (#:dir "amalgam")
```

The compiled "amalgam.rkt" module has a build dependency on
"alamgam-src.rkt" and its required modules, but it has no run-time or
(for uses of "amalgam.rkt") expand-time dependency on
"alamgam-src.rkt" or its required modules within the "amalgam"
directory.

Currently, the main benefit of `#lang compiler/demod` is to reduce the
startup time of a library that is organized into many modules (on the
order of hundreds). For example, the core of the current Rhombus
implementation is 300+ modules; demodularizing reduces the startup
time for a simple `#lang rhombus "Hello world"` by about 25%. That's
not a big effect, and the saved milliseconds are in the noise for a
realistic program. But from the perspective of building and
maintaining the Rhombus implementation, demodularization removes a
temptation to glom code into bigger chunks.

The implementation of `#lang compiler/demod` relies on a small change
to the Racket expander so that a `#%module-begin` macro can produce
compiled code instead of a syntax object. A `#%module-begin` macro can
return compiled code only when the enclosing module is a top-level
module (not a submodule) and when it is being compiled, not when it is
merely being expanded; the new `syntax-local-compiling-module?`
predicate indicates when a compiled result is allowed. In a mode other
than compilation, the `#%module-begin` macro for `compiler/demod`
produces `require`, `provide`, and submodule forms to re-export
bindings from the module that would be flattened in compile mode. This
approach ensures that `comiler/demod` works with tools like DrRacket
and errortrace, which expand a module before compiling/evaluating it;
demodularization only happens in the case of direct compilation
through tools like `raco make` or `raco setup`.

Although `#lang compiler/demod` is intended primarily for libraries,
when the `#:exe` option is specified in the module body, it works for
end-user programs the same as running `raco demod` without `-s` . In
that mode, syntax object and macros are stipped and unused functions
can be dropped. Both `raco demod` and `#:exe` mode now preserve
submodules and provides, however, so `raco demod` can work on a module
that has a `main` submodule or that exports a `main` function.

There are still limitations related to dynamically required modules.
Most things that `raco exe` can handle should be within reach of
future improvements.