BC: repair linklet optimizer bug
Fix a bug in the bytecode optimizer that was exposed by the improved
demodularizer. Unlike the macro expander, the demodularizer can
generate a linklet that has unexported definitions. Unlike the linklet
form of the expander itself, a linklet generated by the demodularizer
can have imports to be optimized away. With that combination, the
index of an unexported defintion was miscalculated.