compiler/demod: add `#:prune-definitions`
Add a new pass to the demosularizer to resolve
`variable-reference-constant?` and related forms, which extends the
reach of the definition pruner. For example, `#lang racket "hello"`
demodularizes with pruning to less than 1/10 of the previous size.
(The initialization of `object%` remains out of reach for the pruner.)

Also, change the preferred form of the `--garbage-collect` flag for
`raco demod` to `--prune-definitions`.