compiler/demod: make library mode work with #:prune-definitions
When preserving syntax and macros, definitions are kept when they are
reachable via syntax objects without using `datum->syntax`.

This imrpovement requires a new `syntax-bound-interned-scope-symbols`
function from the expander. The new function scans a syntax object's
scopes to determine what interned scopes are relevant to looking up
bindings --- so it's not fast, but it's fast enough for
demodularization.