raco pkg create: add `--original` argument
Support recording an original package source in a package's
"info.rkt". The intent is that package sources will be recorded by a
distribution build, and then other package tools can use this
information.

Package installation propagates this information. When an installed
package's source is not a Git URL, but it has an original source
that's a Git URL, then the original source is recorded in the table of
installed packages --- in the same way as when a catalog provides Git
URL. That makes the package with the recorded source eligible for use
with `--clone` (without first pointing to a different catalog), and it
enables detection of other packages that should be converted for the
same source repository.