raco exe: repair ELF update to add a section
When a NOBITS section is misinterpreted as taking up space in the
file, and when that space would overlap with the string-constant
update, then an incorrect update to the virtual-memory size of the
section was planned. That adjustment, in turn, could trigger
disallowed virtual-address shifts to PROGBITS sections that are later
in memory, leading to a "cannot move SHT_PROGBITS section" error.