compiler/demod: repair for Windows
Because module filesystem paths were inconsistently normalized, `#lang
compiler/demod` builds did not work on Windows. Nomralization should
not adjust the case of module paths, tempting as it may be, because
that fights against the module system's conventions. More generally,
use `single-form-path` instead of `normalize-path` to be consistet
with module-path normalization.