expander: repair for reflective access of for-label portal
Accessing portal syntax bound at the label phase should not
instantiate the enclosing module at the label phase, because any
label-phase binding information in the portal syntax object will be
lost (due to a label-phase shift of a label phase). There's not an
unambigious choice when requesting a for-label portal, but
instantiating at phase 0 is more useful.